requirements.txt
 moviepy-1.0.3
 SpeechRecognition-3.10.1
 replicate-0.25.0
 fastapi-0.110.0
 uvicorn-0.29.0